A bigger commit, fixes: - Attack while following - Attack-follow into peace zones - Escaping from under spell radius (effectRange variable in datapack skills) - Raidboss minion respawn, also takes away normal boss minion respawn timers (for performance reasons) - Polearm hit to another floors (didn't happen often but was possible).

+++ L2_GameServer_c5/java/net/sf/l2j/gameserver/model/L2World.java
@@ -636,13 +636,13 @@ public final class L2World
                 int y1 = _object.getY();
                 int z1 = _object.getZ();
                 
-                int dx = x1 - x;
+                long dx = x1 - x;
                 //if (dx > radius || -dx > radius)
                 //  continue;
-                int dy = y1 - y;
+                long dy = y1 - y;
                 //if (dy > radius || -dy > radius)
                 //  continue;
-                int dz = z1 - z;
+                long dz = z1 - z;
                 
                 if (dx*dx + dy*dy + dz*dz < sqRadius)
                     result.add(_object);
